# frontend/openapi.json
{
  "components": {
    "headers": {
...
      "TraceId": {
        "description": "Request trace identifier.",
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "parameters": {
      "CsrfToken": {
        "description": "Token bound to the current opaque session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "schema": {
          "maxLength": 128,
          "minLength": 1,
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Conflict": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "The resource state conflicts with the request.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
      "Forbidden": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Permission, assignment or CSRF validation failed.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Request rejected.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "A valid opaque session cookie is required.",
        "headers": {
          "X-Trace-Id": {
            "$ref": "#/components/headers/TraceId"
          }
        }
      }
    },
    "schemas": {
...
      "Backup": {
        "additionalProperties": false,
        "properties": {
          "checksum_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "type": [
              "string",
              "null"
            ]
          },
          "completed_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "enum": [
              "manual",
              "scheduled",
              "pre_restore",
              "pre_update"
            ]
          },
          "size_bytes": {
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "enum": [
              "creating",
              "ready",
              "failed"
            ]
          }
        },
        "required": [
          "id",
          "instance_id",
          "kind",
          "status",
          "checksum_sha256",
          "size_bytes",
          "created_at",
          "completed_at"
        ],
        "type": "object"
      },
      "BedrockArchiveUploadJobInteraction": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "const": "bedrock_archive_upload"
          },
          "max_bytes": {
            "const": 4294967296
          },
          "method": {
            "const": "POST"
          },
          "path": {
            "pattern": "^/api/v1/servers/[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[1-8][0-9A-Fa-f]{3}-[89AaBb][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}/imports/zip$",
            "type": "string"
          },
          "required_sha256_header": {
            "const": "x-dmx-archive-sha256"
          },
          "version": {
            "maxLength": 128,
            "minLength": 1,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "kind",
          "instance_id",
          "version",
          "method",
          "path",
          "required_sha256_header",
          "max_bytes"
        ],
        "type": "object"
      },
...
      "InstanceRequest": {
        "additionalProperties": false,
        "properties": {
          "instance_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "instance_id"
        ],
        "type": "object"
      },
      "Job": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "interaction": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobInteraction"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string"
          },
          "progress": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer"
          },
          "requested_by": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "waiting_for_user",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ]
          }
        },
        "required": [
          "id",
          "instance_id",
          "kind",
          "state",
          "progress",
          "requested_by",
          "error_code",
          "error_message",
          "created_at",
          "started_at",
          "finished_at",
          "interaction"
        ],
        "type": "object"
      },
      "JobInteraction": {
        "discriminator": {
          "mapping": {
            "bedrock_archive_upload": "#/components/schemas/BedrockArchiveUploadJobInteraction",
            "oauth_device": "#/components/schemas/OauthDeviceJobInteraction"
          },
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/OauthDeviceJobInteraction"
          },
          {
            "$ref": "#/components/schemas/BedrockArchiveUploadJobInteraction"
          }
        ]
      },
...
      "OauthDeviceJobInteraction": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "oauth_device"
          },
          "user_code": {
            "maxLength": 32,
            "minLength": 4,
            "pattern": "^[A-Z0-9-]+$",
            "type": [
              "string",
              "null"
            ]
          },
          "verification_uri": {
            "description": "HTTPS accounts.hytale.com/device URL with an optional validated user_code query.",
            "format": "uri",
            "maxLength": 4096,
            "type": "string"
          }
        },
        "required": [
          "kind",
          "verification_uri",
          "user_code"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "trace_id"
        ],
        "type": "object"
      },
...
      "SuccessResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "Opaque HttpOnly SameSite=Strict session cookie. Secure is mandatory outside loopback development.",
        "in": "cookie",
        "name": "dmx_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-host game server management API. Secret values are write-only.",
    "title": "DmxServerManager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/backups": {
      "get": {
        "operationId": "listBackups",
        "parameters": [
          {
            "in": "query",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Backup"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Backups for the selected instance."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      },
      "post": {
        "operationId": "createBackup",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Persistent backup job accepted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}": {
      "delete": {
        "operationId": "deleteBackup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Backup archive and record removed."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      },
      "get": {
        "operationId": "getBackup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backup"
                }
              }
            },
            "description": "Backup metadata."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}/download": {
      "get": {
        "operationId": "downloadBackup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Verified backup archive.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}/restore": {
      "post": {
        "operationId": "restoreBackup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Persistent restore job accepted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}